fix(docs): put the recomputed caveats where the stale ones were
The permutation caveat now reads the same everywhere it appears: 5 of 28
pairs among the top eight at p < 0.05 two-sided, one (Messi vs Benzema)
surviving Bonferroni, and every appearance names the tested statistic as
the unweighted mean of per-season scores, not the ranking's
minutes-weighted composite. The dashboard's A-vs-B tab is keyed by
player_id so the two Luis Suárezes stay two players. Identity aligns on
93.9% of the 65,069 collapsed player-seasons with 1,240 players
unresolved. index.qmd's gate is twelve requirements, its limitations
bullet stops denying the Champions League and internationals it counts,
its stale freeze is deleted, and its layout and reproduction sections
shrink to pointers at README. The BCa coverage study is named for what
it simulates, PENDING stops calling shipped BCa future work, awards and
defender figures move to the new pipeline's numbers (median 38,
Cannavaro 2,363rd, best defender 170th, top 50 88/12/0), and
method-template.qmd leaves the chapter list for docs/.

+++ gambeta/verdict.py
@@ -98,11 +98,13 @@ def summary(placed: pd.DataFrame, ranking: pd.DataFrame) -> pd.DataFrame:
     total = len(ranking)
     top_slice = max(int(round(0.01 * total)), 1)
 
+    # The median of an even count of winners lands between two ranks. Truncating
+    # it to an int reported "37" for a median of 37.5, so it is kept exact.
     rows = [
         ("award winners in the window", len(placed)),
         ("of those, present in our data", len(found)),
         ("of those, clearing all requirements", len(qualified)),
-        ("median rank of a winner", int(found["rank"].median()) if len(found) else 0),
+        ("median rank of a winner", float(found["rank"].median()) if len(found) else 0.0),
         (f"winners inside our top {top_slice}", int((found["rank"] <= top_slice).sum())),
         ("winners our gate rejects", int(len(found) - len(qualified))),
     ]
